modeloLocal/requirements.txt
fastapi
uvicorn
google-genai
pydantic
python-dotenv
sqlalchemy
transformers
torch
scikit-learn
